+++ Example/SimpleVideo.swift
@@ -1,15 +1,16 @@
 //
-//  SimpleAudioDelegate.swift
+//  SimpleVideo.swift
 //  ModernAVPlayer_Example
 //
-//  Created by ankierman on 16/10/2019.
+//  Created by ankierman on 31/10/2019.
 //  Copyright © 2019 CocoaPods. All rights reserved.
 //
 
+import AVFoundation
 import ModernAVPlayer
 import UIKit
 
-final class SimpleAudioVC: UIViewController {
+final class SimpleVideoVC: UIViewController {
 
     // MARK: - Inputs
 
@@ -19,25 +20,32 @@ final class SimpleAudioVC: UIViewController {
     }()
     private var currentTime: Double?
     private var itemDuration: Double?
-    private let dataSource: [MediaResource] = [.live, .remote, .local, .invalid]
+    private let seekTime: Double = 5
+    private let dataSource: [MediaResource] = [.custom("https://file-examples.com/wp-content/uploads/2018/04/file_example_MOV_480_700kB.mov")]
 
     // MARK: - Interface Buidler
 
     @IBOutlet weak private var stateLabel: UILabel!
     @IBOutlet weak private var timingLabel: UILabel!
-    @IBOutlet weak private var mediaPicker: UIPickerView!
+    @IBOutlet weak private var playerVideo: AVPlayerView!
+    @IBOutlet weak private var prevSeek: UIButton!
+    @IBOutlet weak private var nextSeek: UIButton!
     
     // MARK: - LifeCycle
 
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        title = "Simple Audio"
-        player.delegate = self
-        mediaPicker.dataSource = self
-        mediaPicker.delegate = self
+        title = "Simple HLS Video"
+        prevSeek.setTitle("- \(seekTime)", for: .normal)
+        nextSeek.setTitle("+ \(seekTime)", for: .normal)
 
+        player.delegate = self
         player.load(media: dataSource[0].playerMedia, autostart: false)
+
+        guard let videoLayer = playerVideo.layer as? AVPlayerLayer
+            else { return }
+        videoLayer.player = player.player
     }
 
     // MARK: - Commands
@@ -60,22 +68,23 @@ final class SimpleAudioVC: UIViewController {
     }
 
     @IBAction func prevSeek(_ sender: UIButton) {
-        guard let position = currentTime else { return }
-        let newPosition = position - 15
-        guard newPosition >= 0 else { return }
-        player.seek(position: newPosition)
+        guard let currentPosition = currentTime else { return }
+        let position = max(currentPosition - seekTime, 0)
+        player.seek(position: position)
     }
 
     @IBAction func nextSeek(_ sender: UIButton) {
-        guard let position = currentTime, let duration = itemDuration
+        guard let currentPosition = currentTime, let duration = itemDuration
             else { return }
-        let newPosition = position + 15
-        guard newPosition < duration else { return }
-        player.seek(position: position + 15)
+
+        let position = currentPosition + seekTime
+        guard position < duration else { return }
+
+        player.seek(position: position)
     }
 }
 
-extension SimpleAudioVC: ModernAVPlayerDelegate {
+extension SimpleVideoVC: ModernAVPlayerDelegate {
     func modernAVPlayer(_ player: ModernAVPlayer, didStateChange state: ModernAVPlayer.State) {
         DispatchQueue.main.async { self.stateLabel.text = "State: " + state.description }
     }
@@ -87,26 +96,5 @@ extension SimpleAudioVC: ModernAVPlayerDelegate {
 
     func modernAVPlayer(_ player: ModernAVPlayer, didItemDurationChange itemDuration: Double?) {
         self.itemDuration = itemDuration
-    }
-}
-
-extension SimpleAudioVC: UIPickerViewDataSource {
-    func numberOfComponents(in pickerView: UIPickerView) -> Int {
-        return 1
-    }
-
-    func pickerView(_ pickerView: UIPickerView, numberOfRowsInComponent component: Int) -> Int {
-        return dataSource.count
-    }
-}
-
-extension SimpleAudioVC: UIPickerViewDelegate {
-    func pickerView(_ pickerView: UIPickerView, titleForRow row: Int, forComponent component: Int) -> String? {
-        return dataSource[row].description
-    }
-
-    func pickerView(_ pickerView: UIPickerView, didSelectRow row: Int, inComponent component: Int) {
-        let playerMedia = dataSource[row].playerMedia
-        player.load(media: playerMedia, autostart: false)
     }
 }